Refactor package structure to organize features and improve code maintainability
