+++ app/src/main/java/com/theimpartialai/speechScribe/core/components/navigation/BottomNavigationBar.kt
@@ -1,0 +1,82 @@
+package com.theimpartialai.speechScribe.core.components.navigation
+
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Icon
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.NavigationBarItemDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavGraph.Companion.findStartDestination
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
+import com.theimpartialai.speechScribe.core.ui.theme.Purple50
+import com.theimpartialai.speechScribe.core.ui.theme.Purple60
+import com.theimpartialai.speechScribe.core.ui.theme.White
+
+@Composable
+fun BottomNavigationBar(
+    navController: NavHostController,
+    state: Boolean,
+    modifier: Modifier = Modifier
+) {
+    val screens = listOf(
+        NavigationItem.SavedRecordings,
+        NavigationItem.RecordingScreen,
+        NavigationItem.Settings
+    )
+
+    NavigationBar(
+        modifier = modifier,
+        containerColor = Purple60
+    ) {
+        val navBackStackEntry by navController.currentBackStackEntryAsState()
+        val currentRoute = navBackStackEntry?.destination?.route
+
+        screens.forEach { screen ->
+            NavigationBarItem(
+                label = {
+                    Text(text = screen.label)
+                },
+                icon = {
+                    Icon(
+                        modifier = Modifier.size(28.dp),
+                        imageVector = ImageVector.vectorResource(id = screen.icon),
+                        contentDescription = screen.label
+                    )
+                },
+                selected = currentRoute == screen.route,
+                onClick = {
+                    navController.navigate(screen.route) {
+                        popUpTo(navController.graph.findStartDestination().id) {
+                            saveState = state
+                        }
+                        launchSingleTop = true
+                        restoreState = true
+                    }
+                },
+                colors = NavigationBarItemDefaults.colors(
+                    selectedIconColor = White,
+                    selectedTextColor = White,
+                    indicatorColor = Purple50,
+                    unselectedIconColor = Purple50,
+                    unselectedTextColor = Purple50
+                )
+            )
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun BottomNavigationBarPreview() {
+    BottomNavigationBar(navController = rememberNavController(), state = true)
+}
+
