--- app/src/main/java/com/theimpartialai/speechScribe/ui/theme/Color.kt
@@ -1,20 +1,0 @@
-package com.theimpartialai.speechScribe.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Purple800 = Color(0xFF2b3056)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val Purple50 = Color(0xFF423a73)
-val Purple60 = Color(0xFF161531)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
-
-val White = Color(0xFFFFFFFF)
-val LightGray = Color(0xFFe4e5e9)
-
-val Blue = Color(0xFF1e1645)
-val DarkBlue = Color(0xFF0d0b21)
